requirements-test.txt
-r requirements.txt
flake8
# pytest will be brought in by pytest-cov
pytest-cov
testfixtures
Sphinx==1.5.3
pyyaml
requests